Grouped steps by object using folders
